Detect current mplayer revisions even if they were svn export-ed
If mplayer was built from "svn export"-ed sources, it doesn't know
its svn revision number, causing version detection to fail, and
phonon-mplayer to assume that mplayer is prehistoric.

Take a guess based on the copyright year, if it's from 2010 or
newer, it can't be as prehistoric as assumed.

While at it, mark the static QRegExps used for version detection
static